requirement.txt
numpy
matplotlib
torch
torchvision
segmentation-models-pytorch
Pillow
tqdm
request
